+++ aftiktuna-rust/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Save format version 3.0, making it incompatible with previous save files
 - Expanded shopkeeper definition to allow setting a custom price
+- Shops can now have a limited quantity of items in stock
 
 ### Contributors for this release